feat: oscillator and periodic
