+++ packages/react-native-audio-api/src/core/index.ts
@@ -1,7 +1,14 @@
+export { default as AnalyserNode } from './analysis/AnalyserNode';
+
 export { default as AudioBuffer } from './AudioBuffer';
 export { default as AudioNode } from './AudioNode';
 export { default as AudioParam } from './AudioParam';
+export { default as PeriodicWave } from './PeriodicWave';
+
 export { default as AudioDestinationNode } from './destinations/AudioDestinationNode';
+
 export { default as BiquadFilterNode } from './effects/BiquadFilterNode';
 export { default as GainNode } from './effects/GainNode';
 export { default as StereoPannerNode } from './effects/StereoPannerNode';
+
+export { default as AudioScheduledSourceNode } from './sources/AudioScheduledSourceNode';
