+++ packages/react-native-audio-api/src/types/internal/IPeriodicWave.ts
@@ -1,0 +1,1 @@
+export default interface IPeriodicWave {}
